fix: address review comments on MCP server example
- Factor NDJSON transport helpers into shared tests/helpers.js module
  to eliminate duplicated reader logic across all 4 test files (#1)
- Remove confusing BBP parenthetical comment; the implementation uses
  Machin's formula throughout (#9)
- Replace Math.random() with seeded PRNG (mulberry32) in test code
  constants for deterministic, non-flaky test results (#10)
- Tighten timing assertion from executeMs > 0 to > 1 for meaningful
  validation (#10)

Note: Issues #2-8 and #11 from the review were already addressed in
the current code (setScratchSize, safe JSON.stringify, shell script
portability, PowerShell cleanup, timing log correctness, u32 bounds
check, Node >= 20 in README + engines, correct JSDoc path).
