+++ src/js-host-api/examples/mcp-server/tests/config.test.js
@@ -14,170 +14,7 @@
 // ─────────────────────────────────────────────────────────────────────
 
 import { describe, it, expect, beforeAll, afterAll } from 'vitest';
-import { spawn } from 'node:child_process';
-import { fileURLToPath } from 'node:url';
-import { dirname, join } from 'node:path';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const SERVER_PATH = join(__dirname, '..', 'server.js');
-const PROTOCOL_VERSION = '2025-11-25';
-
-// ── NDJSON Utilities ────────────────────────────────────────────────
-
-function send(proc, message) {
-    proc.stdin.write(JSON.stringify(message) + '\n');
-}
-
-/**
- * Shared per-process stdout state to correctly handle NDJSON streams.
- * Ensures that multiple messages in a single chunk are not dropped and
- * that leftover buffered data is preserved across waitForResponse calls.
- */
-const stdoutState = new WeakMap();
-
-function getStdoutState(proc) {
-    let state = stdoutState.get(proc);
-    if (state) return state;
-
-    state = {
-        buffer: '',
-        pendingLines: [],
-        waiters: [],
-        listening: false,
-        listener: null,
-    };
-
-    const listener = (chunk) => {
-        state.buffer += chunk.toString();
-
-        while (true) {
-            const idx = state.buffer.indexOf('\n');
-            if (idx === -1) break;
-
-            let line = state.buffer.slice(0, idx);
-            state.buffer = state.buffer.slice(idx + 1);
-
-            // Normalize Windows line endings and skip empty lines.
-            line = line.replace(/\r$/, '');
-            if (line.length === 0) continue;
-
-            if (state.waiters.length > 0) {
-                const { resolve, reject } = state.waiters.shift();
-                try {
-                    resolve(JSON.parse(line));
-                } catch (_err) {
-                    reject(new Error(`Invalid JSON from server: ${line}`));
-                }
-            } else {
-                state.pendingLines.push(line);
-            }
-        }
-    };
-
-    state.listener = listener;
-    stdoutState.set(proc, state);
-    return state;
-}
-
-function waitForResponse(proc) {
-    const state = getStdoutState(proc);
-
-    return new Promise((resolve, reject) => {
-        const deliverFromLine = (line) => {
-            try {
-                resolve(JSON.parse(line));
-            } catch (_err) {
-                reject(new Error(`Invalid JSON from server: ${line}`));
-            }
-        };
-
-        // If we already have a complete line buffered, use it immediately.
-        if (state.pendingLines.length > 0) {
-            const line = state.pendingLines.shift();
-            deliverFromLine(line);
-            return;
-        }
-
-        // Otherwise, enqueue this waiter and let the shared listener fulfill it.
-        state.waiters.push({ resolve, reject });
-
-        // Attach the shared listener once per process.
-        if (!state.listening) {
-            proc.stdout.on('data', state.listener);
-            state.listening = true;
-        }
-    });
-}
-
-/**
- * Spawn a server with the given env overrides, perform the MCP
- * handshake, and return { server, messageId, callTool, listTools }.
- *
- * @param {Record<string, string>} envOverrides
- * @returns {Promise<{server: import('node:child_process').ChildProcess, messageId: {value: number}, callTool: (code: string) => Promise<object>, listTools: () => Promise<object>}>}
- */
-async function spawnServer(envOverrides = {}) {
-    const messageId = { value: 1 };
-    const server = spawn('node', [SERVER_PATH], {
-        stdio: ['pipe', 'pipe', 'pipe'],
-        env: { ...process.env, ...envOverrides },
-    });
-
-    // Collect stderr for debugging (vitest captures it)
-    const stderrChunks = [];
-    server.stderr.on('data', (d) => {
-        stderrChunks.push(d.toString());
-        process.stderr.write(`[config-test] ${d}`);
-    });
-
-    // MCP handshake — initialize
-    send(server, {
-        jsonrpc: '2.0',
-        id: messageId.value++,
-        method: 'initialize',
-        params: {
-            protocolVersion: PROTOCOL_VERSION,
-            capabilities: {},
-            clientInfo: { name: 'vitest-config-client', version: '1.0.0' },
-        },
-    });
-
-    const initResponse = await waitForResponse(server);
-    expect(initResponse.result).toBeDefined();
-
-    // MCP handshake — initialized notification
-    send(server, {
-        jsonrpc: '2.0',
-        method: 'notifications/initialized',
-    });
-    await new Promise((r) => setTimeout(r, 200));
-
-    /** Call execute_javascript and return the full response. */
-    const callTool = async (code) => {
-        send(server, {
-            jsonrpc: '2.0',
-            id: messageId.value++,
-            method: 'tools/call',
-            params: {
-                name: 'execute_javascript',
-                arguments: { code },
-            },
-        });
-        return waitForResponse(server);
-    };
-
-    /** Call tools/list and return the full response. */
-    const listTools = async () => {
-        send(server, {
-            jsonrpc: '2.0',
-            id: messageId.value++,
-            method: 'tools/list',
-        });
-        return waitForResponse(server);
-    };
-
-    return { server, messageId, callTool, listTools, stderrChunks };
-}
+import { spawnServer } from './helpers.js';
 
 // ── Custom CPU Timeout ──────────────────────────────────────────────
