+++ src/js-host-api/examples/mcp-server/tests/mcp-server.test.js
@@ -10,105 +10,7 @@
 
 import { describe, it, expect, beforeAll, afterAll } from 'vitest';
 import { spawn } from 'node:child_process';
-import { fileURLToPath } from 'node:url';
-import { dirname, join } from 'node:path';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-/** Path to server.js — one directory up from tests/ */
-const SERVER_PATH = join(__dirname, '..', 'server.js');
-
-/**
- * The protocol version the MCP SDK (v1.26.0) expects.
- * Must match LATEST_PROTOCOL_VERSION in the SDK.
- */
-const PROTOCOL_VERSION = '2025-11-25';
-
-// ── NDJSON Framing ──────────────────────────────────────────────────
-//
-// MCP stdio transport uses newline-delimited JSON (NDJSON):
-//   - Send: JSON.stringify(message) + '\n'
-//   - Receive: read lines, parse each as JSON
-//
-// NOT Content-Length / LSP framing.
-
-/**
- * Send a JSON-RPC message to the server via stdin (NDJSON framing).
- *
- * @param {import('node:child_process').ChildProcess} proc
- * @param {object} message — JSON-RPC message object
- */
-function send(proc, message) {
-    proc.stdin.write(JSON.stringify(message) + '\n');
-}
-
-// Shared per-process line reader state: buffer, queued lines, and waiters.
-// Uses a WeakMap so the reader survives across multiple waitForResponse calls
-// without dropping lines that arrived in the same stdout chunk.
-const procLineState = new WeakMap();
-
-function ensureLineReader(proc) {
-    let state = procLineState.get(proc);
-    if (state) return state;
-
-    state = {
-        buffer: '',
-        lines: [],
-        waiters: [],
-    };
-
-    const onData = (chunk) => {
-        state.buffer += chunk.toString();
-        let idx;
-        while ((idx = state.buffer.indexOf('\n')) !== -1) {
-            let line = state.buffer.slice(0, idx).replace(/\r$/, '');
-            state.buffer = state.buffer.slice(idx + 1);
-            if (line.length === 0) {
-                continue;
-            }
-
-            if (state.waiters.length > 0) {
-                const { resolve, reject } = state.waiters.shift();
-                try {
-                    resolve(JSON.parse(line));
-                } catch (_err) {
-                    reject(new Error(`Invalid JSON from server: ${line}`));
-                }
-            } else {
-                state.lines.push(line);
-            }
-        }
-    };
-
-    proc.stdout.on('data', onData);
-    procLineState.set(proc, state);
-    return state;
-}
-
-/**
- * Wait for the next JSON-RPC response from the server's stdout.
- * Reads newline-delimited JSON via a persistent per-process line buffer.
- *
- * @param {import('node:child_process').ChildProcess} proc
- * @returns {Promise<object>} — parsed JSON-RPC response
- */
-function waitForResponse(proc) {
-    return new Promise((resolve, reject) => {
-        const state = ensureLineReader(proc);
-
-        if (state.lines.length > 0) {
-            const line = state.lines.shift();
-            try {
-                resolve(JSON.parse(line));
-            } catch (_err) {
-                reject(new Error(`Invalid JSON from server: ${line}`));
-            }
-            return;
-        }
-
-        state.waiters.push({ resolve, reject });
-    });
-}
+import { send, waitForResponse, SERVER_PATH, PROTOCOL_VERSION } from './helpers.js';
 
 // ── Test Suite ──────────────────────────────────────────────────────
